+++ src/gtt_application_window.c
@@ -34,9 +34,9 @@
 #include "gtt_current_project.h"
 #include "gtt.h"
 #include "gtt_log.h"
-#include "menucmd.h"
+#include "gtt_menu_commands.h"
 #include "gtt_menus.h"
-#include "notes-area.h"
+#include "gtt_notes_area.h"
 #include "gtt_preferences.h"
 #include "gtt_props_dlg_project.h"
 #include "gtt_timer.h"
