+++ src/gtt_menus.c
@@ -28,7 +28,7 @@
 #include "gtt_export.h"
 #include "gtt.h"
 #include "gtt_journal.h"
-#include "menucmd.h"
+#include "gtt_menu_commands.h"
 #include "gtt_plug_in.h"
 #include "gtt_timer.h"
