+++ src/gtt_notes_area.h
