+++ src/gtt_project_queries.c
@@ -18,11 +18,12 @@
 
 #include "config.h"
 
+#include "gtt_project_queries.h"
+
 #include <glib.h>
 
 #include "gtt_current_project.h"
 #include "gtt_project.h"
-#include "proj-query.h"
 
 /* =========================================================== */
