+++ src/gtt_project_queries.h
