+++ src/gtt_toolbar.c
@@ -28,7 +28,7 @@
 #include "gtt_help_popup.h"
 #include "gtt.h"
 #include "gtt_journal.h"
-#include "menucmd.h"
+#include "gtt_menu_commands.h"
 #include "gtt_menus.h"
 #include "gtt_myoaf.h"
 #include "gtt_preferences.h"
